+++ tests/Alarm_mockups.cpp
@@ -1,11 +1,8 @@
 #include "Alarm_mockups.h"
 
-
 HALbool lamp(set_lamp);
+MockupBuzzerManager buzzer;
 bool lamp_status;
-bool buzzer_status;
-unsigned int buzzer_freq;
-unsigned long buzzer_duration;
 bool EEPROM_write;
 bool activated;
 bool stopped;
@@ -16,25 +13,12 @@ void writeEEPROM() { EEPROM_write = true; }
 void activation_callback() { activated = true; }
 void stop_callback() { stopped = true; }
 
-void buzzerTone(unsigned int freq, unsigned long duration)
-{
-    buzzer_status = true;
-    buzzer_freq = freq;
-    buzzer_duration = duration;
-}
-
-void buzzerNoTone()
-{
-    buzzer_status = false;
-}
 
 void reset_alarm_mockups()
 {
     lamp = HALbool(set_lamp);
+    buzzer = MockupBuzzerManager();
     lamp_status = false;
-    buzzer_status = false;
-    buzzer_freq = 0;
-    buzzer_duration = 0;
     EEPROM_write = false;
     activated = false;
     stopped = false;