+++ tests/Alarm_mockups.h
@@ -8,13 +8,22 @@
 #endif
 
 #include "src/AlarmClock/HALbool.h"
+#include "src/AlarmClock/BuzzerManager.h"
+
+class MockupBuzzerManager : public BuzzerManager
+{
+public:
+    MockupBuzzerManager() : BuzzerManager(255) { };
+    bool get_status() const { return status_; };
+    BuzzerTone get_tone() const { return tone_; };
+
+};
 
 extern HALbool lamp;
+extern MockupBuzzerManager buzzer;
 extern bool lamp_status;
 
 extern bool buzzer_status;
-extern unsigned int buzzer_freq;
-extern unsigned long buzzer_duration;
 
 extern bool EEPROM_write;
 
@@ -24,9 +33,6 @@ extern bool stopped;
 void reset_alarm_mockups();
 
 void set_lamp(bool status);
-
-void buzzerTone(unsigned int freq, unsigned long duration = 0);
-void buzzerNoTone();
 
 void writeEEPROM();